agri-link-backend/requirements.txt
alembic==1.14.1
aniso8601==10.0.1
bcrypt==5.0.0
blinker==1.8.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.44
SQLAlchemy-serializer==1.4.12
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2

# Production server
gunicorn==23.0.0
psycopg2-binary==2.9.9

# AI (optional)
openai==1.60.0
